#66 Rename StaticDLL to SharedDLL. #67 Merge Editor Into Game. #68 Clean up settings, states, menus
Cleaned up alot of code.
Refactored so my dll was neater and didn't contain uneeded classes that
game would use.
Joined editor and game together. (Better this way)
Restructured settings to have basesettings class.(Made a kind of map of
structs to store different types for the settings).
Removed the girl image and drew a ghost for now as a placeholder.
Attempted to clean up gitignore.(Not very well)